+++ controller/book-controller.js
@@ -31,12 +31,29 @@ exports.getSingleBookById = async (req, res) => {
     });
 };
 
+exports.getSingleBookByName = async (req, res) => {
+    const { name } = req.params;
+    const book = await bookModel.findOne({
+        name:name,
+    });
+    if (!book) {
+        return res.status(404).json({
+            success: false,
+            message: "Book not found"
+        });
+    };
+    return res.status(200).json({
+        success: true,
+        message: book
+    });
+};
+
 exports.getAllIssuedBooks = async (req, res) => {
     const users = await UserModel.find({
         issuedBook: { $exists: true }
     }).populate("issuedBook");
 
-    const issuedBooks = users.map((each)=>new IssuedBook(each));
+    const issuedBooks = users.map((each) => new IssuedBook(each));
 
     if (issuedBooks.length === 0)
         return res.status(404).json({
@@ -48,3 +65,43 @@ exports.getAllIssuedBooks = async (req, res) => {
         data: issuedBooks,
     });
 };
+
+exports.AddNewBook = async (req, res) => {
+    const { data } = req.body;
+
+    if (!data) {
+        return res.status(400).json({
+            success: false,
+            message: "No data provided",
+        });
+    }
+
+    await bookModel.create(data);
+
+    const allBooks = await BookModel.find();
+
+    return res.status(201).json({
+        success: true,
+        data: allBooks,
+    });
+};
+
+exports.updateBookById = async (req, res) => {
+    const { id } = req.params;
+    const { data } = req.body;
+  
+    const updatedBook = await BookModel.findOneAndUpdate(
+      {
+        _id: id,
+      },
+      data,
+      {
+        new: true,
+      }
+    );
+  
+    return res.status(200).json({
+      success: true,
+      data: updatedBook,
+    });
+  };
