+++ models/user-model.js
@@ -20,6 +20,10 @@ const userSchema = new Schema(
             ref: "Book",
             required: false,
         },
+        issuedDate: {
+            type: String,
+            required: false,
+        },
         returnDate: {
             type: String,
             required: false,
