+++ routes/users.js
@@ -1,4 +1,5 @@
 const express = require("express");
+const { getAllUsers, getSingleUserById, deleteUser, updateUserById, createUser, getSubscriptionDetailsById, createNewUser } = require("../controller/user-controller");
 const { users } = require('../data/users.json');
 
 const router = express.Router();
@@ -9,9 +10,7 @@ const router = express.Router();
  * Access: Public
  * Parameters: none
 */
-router.get('/', (req, res) => {
-    res.status(200).json({ success: true, data: users, });
-});
+router.get('/', getAllUsers );
 
 /** 
  * Route: /users/:id
@@ -21,20 +20,7 @@ router.get('/', (req, res) => {
  * Parameters: id
 */
 
-router.get('/:id', (req, res) => {
-    const { id } = req.params;
-    const user = users.find((each) => each.id === id);
-    if (!user) {
-        return res.status(404).json({
-            success: false,
-            message: "User not found !",
-        });
-    }
-    return res.status(200).json({
-        success: true,
-        message: user,
-    });
-});
+router.get('/:id', getSingleUserById);
 
 
 /** 
@@ -45,29 +31,7 @@ router.get('/:id', (req, res) => {
  * Parameters: none
 */
 
-router.post('/', (req, res) => {
-    const { id, name, surname, email, subscriptionType, subscriptionDate } = req.body;
-
-    const user = users.find((each) => each.id === id);
-    if (user) {
-        return res.status(404).json({
-            success: false,
-            message: 'User exists with this id',
-        });
-    }
-    users.push({
-        id,
-        name,
-        surname,
-        email,
-        subscriptionType,
-        subscriptionDate,
-    });
-    return res.status(201).json({
-        success: true,
-        message: users,
-    });
-});
+router.post('/', createNewUser);
 
 /** 
  * Route: /users/:id
@@ -77,27 +41,7 @@ router.post('/', (req, res) => {
  * Parameters: id
 */
 
-router.put('/:id', (req, res) => {
-    const { id } = req.params;
-    const { data } = req.body;
-
-    const user = users.find((each) => each.id === id);
-    if (!user) return res.status(404).json({ success: false, message: "User not found !" });
-
-    const updatedUser = users.map((each) => {
-        if (each.id === id) {
-            return {
-                ...each,
-                ...data,
-            };
-        }
-        return each;
-    });
-    return res.status(200).json({
-        success: true,
-        data: updatedUser,
-    })
-});
+router.put('/:id', updateUserById);
 
 /** 
  * Route: /users/:id
@@ -107,23 +51,7 @@ router.put('/:id', (req, res) => {
  * Parameters: id
 */
 
-router.delete("/:id", (req, res) => {
-    const { id } = req.params;
-    const user = users.find((each) => each.id === id);
-
-    if (!user) {
-        return res.status(404).json({
-            success: false,
-            message: "User to be delete was not found!"
-        });
-    }
-    const index = users.indexOf(user);
-    users.splice(index, 1);
-    return res.status(202).json({
-        success: true,
-        data: users
-    })
-});
+router.delete("/:id", deleteUser);
 
 /**
  * Route: /users/subscription-details/:id
@@ -132,72 +60,6 @@ router.delete("/:id", (req, res) => {
  * Access: Public
  * Parameters: id
  */
- router.get("/subscription-details/:id", (req, res) => {
-    const { id } = req.params;
-  
-    const user = users.find((each) => each.id === id);
-  
-    if (!user)
-      return res.status(404).json({
-        success: false,
-        message: "User not found",
-      });
-  
-    const getDateInDays = (data = "") => {
-      let date;
-      if (data === "") {
-        // current date
-        date = new Date();
-      } else {
-        // getting date on basis of data variable
-        date = new Date(data);
-      }
-      let days = Math.floor(date / (1000 * 60 * 60 * 24));
-      return days;
-    };
-  
-    const subscriptionType = (date) => {
-      if (user.subscriptionType === "Basic") {
-        date = date + 90;
-      } else if (user.subscriptionType === "Standard") {
-        date = date + 180;
-      } else if (user.subscriptionType === "Premium") {
-        date = date + 365;
-      }
-      return date;
-    };
-  
-    // Subscription expiration calculation
-    // January 1, 1970, UTC. // milliseconds
-    let returnDate = getDateInDays(user.returnDate);
-    let currentDate = getDateInDays();
-    let subscriptionDate = getDateInDays(user.subscriptionDate);
-    let subscriptionExpiration = subscriptionType(subscriptionDate);
-  
-    console.log("Return Date ", returnDate);
-    console.log("Current Date ", currentDate);
-    console.log("Subscription Date ", subscriptionDate);
-    console.log("Subscription expiry date", subscriptionExpiration);
-  
-    const data = {
-      ...user,
-      subscriptionExpired: subscriptionExpiration < currentDate,
-      daysLeftForExpiration:
-        subscriptionExpiration <= currentDate
-          ? 0
-          : subscriptionExpiration - currentDate,
-      fine:
-        returnDate < currentDate
-          ? subscriptionExpiration <= currentDate
-            ? 200
-            : 100
-          : 0,
-    };
-  
-    res.status(200).json({
-      success: true,
-      data,
-    });
-  });
+ router.get("/subscription-details/:id", getSubscriptionDetailsById);
 
 module.exports = router;
